+++ package.json
@@ -13,7 +13,13 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
+    "body-parser": "^1.20.0",
+    "cors": "^2.8.5",
+    "express": "^4.17.3",
+    "fast-csv": "^4.3.6",
     "http-proxy-middleware": "^2.0.4",
+    "multer": "^1.4.4",
+    "mysql2": "^2.3.3",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
@@ -23,6 +29,7 @@
     "react-virtualized": "^9.22.3",
     "reactdom": "^2.0.0",
     "reactstrap": "^9.0.2",
+    "sequelize": "^6.19.0",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4",
     "xlsx": "^0.18.5"
