+++ src/components/Manage_UploadStudent.js
@@ -2,7 +2,7 @@ import * as React from 'react';
 import './Manage.css';
 import SideBar from '../components/views/Manage_Header';
 import styled from "styled-components";
-import TimetableBox from './views/TimetableBox2'
+import TimetableBox from './views/TimetableBox'
 import * as BsIcons from 'react-icons/bs';
 import { Container } from '@mui/material';
 import { AiOutlineUser, AiTwotoneHome } from "react-icons/ai";
